Added files via upload
This is an example of the VPPM modulation scheme. User can change the modulation scheme from here. An output files of SNR,BER and goodput will be generated. In addition the routing table of the implemented example should also be generated when running MyFirst.cc using the NS3 WAF build system.